request setup is now 3 files
